Split EmailDisplay And Inbox Into Separate Files
