+++ frontend/src/components/inbox/inbox.jsx
@@ -1,10 +1,7 @@
-import { Readability } from "@mozilla/readability";
+import EmailDisplay from "./emailDisplay";
 import { useEffect, useRef, useState } from "react";
-import ReactDom from "react-dom";
 import ArrowIcon from "../../assets/InboxArrow";
-import ReaderViewIcon from "../../assets/ReaderView";
 import { emailsPerPage } from "../../assets/constants";
-import "./emailDisplay.css";
 import "./emailEntry.css";
 import "./emailList.css";
 
@@ -108,75 +105,6 @@ function InboxEmailList({ displaySummaries, emailList, curEmail, onClick }) {
         {emails()}
       </div>
     </div>
-  );
-}
-
-function EmailDisplay({ curEmail }) {
-  const date = getDate(curEmail.received_at);
-  return (
-    <div className="email-display">
-      <div className="header">
-        <div className="from">{curEmail.sender}</div>
-        <div className="title">{curEmail.subject}</div>
-        <div className="to">{`To: ${curEmail.recipients}`}</div>
-        <div className="date">{`Date: ${date}`}</div>
-        <div className="reader-view">
-          <ReaderView curEmail={curEmail} />
-        </div>
-      </div>
-      <div className="body">
-        <div className="content-container">
-          <div className="content">{curEmail.body}</div>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-function ReaderView({ curEmail }) {
-  const [text, setText] = useState("Loading ...");
-  const [displaying, setDisplaying] = useState(false);
-
-  function displayReaderView() {
-    if (!displaying) {
-      const parser = new DOMParser();
-      const doc = parser.parseFromString(curEmail.body, "text/html");
-      const article = new Readability(doc).parse();
-      setText(article.textContent);
-    }
-    setDisplaying(!displaying);
-  }
-
-  useEffect(() => {
-    setDisplaying(false);
-    setText("Loading...");
-  }, [curEmail]); // Inefficient way to clean state when email switches
-
-  return (
-    <div>
-      <div className="icon-container" onClick={displayReaderView}>
-        <ReaderViewIcon />
-      </div>
-      <PopUp isOpen={displaying} handleClose={displayReaderView}>
-        <div>{text}</div>
-      </PopUp>
-    </div>
-  );
-}
-
-function PopUp({ isOpen, handleClose, children }) {
-  if (!isOpen) return null;
-  return ReactDom.createPortal(
-    <>
-      <div className="overlay-background" />
-      <div className="pop-up-container">
-        {children}
-        <div className="button" onClick={handleClose}>
-          Click To Close
-        </div>
-      </div>
-    </>,
-    document.getElementById("portal")
   );
 }
